--- Assets/Animations/Unity Animation/CrossFade_PauseMenuEntry.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0e7468fbd0a442043b6d7c1986f4ab87
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
