--- Assets/Animations/Unity Animation/CrossFade_PauseMenuFadeIn.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c4d3b61e81c25b049b95e42ef20c9cef
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
