+++ Assets/LevelControl.cs
@@ -2,12 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
-//using UnityEngine.UI;
 
 public class LevelControl : MonoBehaviour
 {
-    public Animator crossFadeObj;
-    //public Slider progressBar;
+    public LevelLoader levelLoadr;
 
     // Start is called before the first frame update
     void Start()
@@ -25,42 +23,11 @@ public class LevelControl : MonoBehaviour
     {
         if(other.tag=="Player")
         {
-            LoadNextLevel();            
+            levelLoadr.LoadNextLevel();            
         }
         else
         {
 
         }
     }
-
-    public void LoadNextLevel()
-    {
-        StartCoroutine(LoadLevel(SceneManager.GetActiveScene().buildIndex+1));
-    }
-
-    IEnumerator LoadLevel(int i)
-    {
-        crossFadeObj.SetTrigger("StartCrossFade");
-        yield return new WaitForSeconds(1);
-        SceneManager.LoadScene(i);
-    }
-
-    //Code To Be Used Later in Life
-    /* IEnumerator LoadAsyncronously(int i)
-    {
-        AsyncOperation operation = SceneManager.LoadSceneAsync(i);
-
-        crossFadeObj.SetTrigger("PlayCrossFadeAnimation");
-
-        while(!operation.isDone)
-        {
-            float progress = Mathf.Clamp01(operation.progress / 0.9f);
-
-            progressBar.value = progress;
-
-            yield return null;
-        }
-
-        crossFadeObj.SetTrigger("StartCrossFade");
-    } */
 }
