+++ Assets/LevelLoader.cs
@@ -1,0 +1,42 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+//using UnityEngine.UI;
+
+public class LevelLoader : MonoBehaviour
+{
+    public Animator crossFadeObj;
+    //public Slider progressBar;
+
+    public void LoadNextLevel()
+    {
+        StartCoroutine(LoadLevel(SceneManager.GetActiveScene().buildIndex+1));
+    }
+
+    IEnumerator LoadLevel(int i)
+    {
+        crossFadeObj.SetTrigger("StartCrossFade");
+        yield return new WaitForSeconds(1);
+        SceneManager.LoadScene(i);
+    }
+
+    //Code To Be Used Later in Life
+    /* IEnumerator LoadAsyncronously(int i)
+    {
+        AsyncOperation operation = SceneManager.LoadSceneAsync(i);
+
+        crossFadeObj.SetTrigger("PlayCrossFadeAnimation");
+
+        while(!operation.isDone)
+        {
+            float progress = Mathf.Clamp01(operation.progress / 0.9f);
+
+            progressBar.value = progress;
+
+            yield return null;
+        }
+
+        crossFadeObj.SetTrigger("StartCrossFade");
+    } */
+}
