+++ Assets/Prefabs/LevelLoader.prefab
@@ -88,6 +88,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1973275969405038797}
+  - component: {fileID: 3884310851339474700}
   m_Layer: 5
   m_Name: LevelLoader
   m_TagString: Untagged
@@ -110,6 +111,19 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &3884310851339474700
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1973275969405038798}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 921b27fae80fc364e945de86374dbfad, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  crossFadeObj: {fileID: 4269776107428303457}
 --- !u!1 &1973275969536015043
 GameObject:
   m_ObjectHideFlags: 0
@@ -211,7 +225,7 @@ GameObject:
   - component: {fileID: 1973275970531620182}
   - component: {fileID: 1973275970531620183}
   - component: {fileID: 1973275970531620176}
-  - component: {fileID: 1973275970531620181}
+  - component: {fileID: 4269776107428303457}
   m_Layer: 5
   m_Name: CrossFade
   m_TagString: Untagged
@@ -258,7 +272,7 @@ Canvas:
   m_SortingBucketNormalizedSize: 0
   m_AdditionalShaderChannelsFlag: 0
   m_SortingLayerID: 0
-  m_SortingOrder: 1
+  m_SortingOrder: 32767
   m_TargetDisplay: 0
 --- !u!114 &1973275970531620183
 MonoBehaviour:
@@ -300,7 +314,7 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
---- !u!95 &1973275970531620181
+--- !u!95 &4269776107428303457
 Animator:
   serializedVersion: 3
   m_ObjectHideFlags: 0
